chore: refactoring according to feedbacks
