+++ package.json
@@ -20,7 +20,7 @@
     "test:integration-legacy": "cucumber-js --fail-fast",
     "test:integration": "jest --config jest.config.integ.js --passWithNoTests",
     "test:protocols": "yarn build:protocols && lerna run test --scope '@aws-sdk/aws-*'",
-    "test:e2e": "node ./scripts/e2e-test/index.js",
+    "test:e2e": "node ./tests/e2e/index.js",
     "local-publish": "node ./scripts/verdaccio-publish/index.js"
   },
   "repository": {
